SubTaskList: move all tests to components/__tests__ folder
